+++ arviz/plots/pairplot.py
@@ -5,14 +5,11 @@ from typing import List, Optional, Union
 import numpy as np
 
 from ..data import convert_to_dataset
+from ..labels import BaseLabeller
+from ..sel_utils import xarray_to_ndarray, xarray_var_iter
+from .plot_utils import get_plotting_function
 from ..rcparams import rcParams
 from ..utils import _var_names, get_coords
-from .plot_utils import (
-    get_plotting_function,
-    xarray_to_ndarray,
-    xarray_var_iter,
-    make_label,
-)
 
 
 def plot_pair(
@@ -31,6 +28,7 @@ def plot_pair(
     fill_last=False,
     divergences=False,
     colorbar=False,
+    labeller=None,
     ax=None,
     divergences_kwargs=None,
     scatter_kwargs=None,
@@ -91,6 +89,9 @@ def plot_pair(
     colorbar: bool
         If True a colorbar will be included as part of the plot (Defaults to False).
         Only works when kind=hexbin
+    labeller : labeller instance, optional
+        Class providing the method `make_label_vert` to generate the labels in the plot.
+        Read the :ref:`label_guide` for more details and usage examples.
     ax: axes, optional
         Matplotlib axes or bokeh figures.
     divergences_kwargs: dicts, optional
@@ -191,13 +192,18 @@ def plot_pair(
     if coords is None:
         coords = {}
 
+    if labeller is None:
+        labeller = BaseLabeller()
+
     # Get posterior draws and combine chains
     dataset = convert_to_dataset(data, group=group)
     var_names = _var_names(var_names, dataset, filter_vars)
     plotters = list(
         xarray_var_iter(get_coords(dataset, coords), var_names=var_names, combined=True)
     )
-    flat_var_names = [make_label(var_name, selection) for var_name, selection, _ in plotters]
+    flat_var_names = [
+        labeller.make_label_vert(var_name, sel, isel) for var_name, sel, isel, _ in plotters
+    ]
 
     divergent_data = None
     diverging_mask = None