+++ arviz/plots/rankplot.py
@@ -4,10 +4,12 @@ from itertools import cycle
 import matplotlib.pyplot as plt
 
 from ..data import convert_to_dataset
+from ..labels import BaseLabeller
+from ..sel_utils import xarray_var_iter
 from ..rcparams import rcParams
 from ..stats.density_utils import _sturges_formula
 from ..utils import _var_names
-from .plot_utils import default_grid, filter_plotters_list, get_plotting_function, xarray_var_iter
+from .plot_utils import default_grid, filter_plotters_list, get_plotting_function
 
 
 def plot_rank(
@@ -21,6 +23,7 @@ def plot_rank(
     colors="cycle",
     ref_line=True,
     labels=True,
+    labeller=None,
     grid=None,
     figsize=None,
     ax=None,
@@ -78,6 +81,9 @@ def plot_rank(
         Whether to include a dashed line showing where a uniform distribution would lie
     labels: bool
         whether to plot or not the x and y labels, defaults to True
+    labeller : labeller instance, optional
+        Class providing the method `make_label_vert` to generate the labels in the plot titles.
+        Read the :ref:`label_guide` for more details and usage examples.
     grid : tuple
         Number of rows and columns. Defaults to None, the rows and columns are
         automatically inferred.
@@ -163,6 +169,9 @@ def plot_rank(
     if bins is None:
         bins = _sturges_formula(posterior_data, mult=2)
 
+    if labeller is None:
+        labeller = BaseLabeller()
+
     rows, cols = default_grid(length_plotters, grid=grid)
 
     chains = len(posterior_data.chain)
@@ -188,6 +197,7 @@ def plot_rank(
         colors=colors,
         ref_line=ref_line,
         labels=labels,
+        labeller=labeller,
         ref_line_kwargs=ref_line_kwargs,
         bar_kwargs=bar_kwargs,
         vlines_kwargs=vlines_kwargs,