+++ arviz/tests/base_tests/test_diagnostics.py
@@ -8,9 +8,9 @@ import pytest
 from numpy.testing import assert_almost_equal
 
 from ...data import from_cmdstan, load_arviz_data
-from ...plots.plot_utils import xarray_var_iter
 from ...rcparams import rc_context, rcParams
 from ...stats import bfmi, ess, mcse, rhat
+from ...sel_utils import xarray_var_iter
 from ...stats.diagnostics import (
     _ess,
     _ess_quantile,
@@ -148,7 +148,7 @@ class TestDiagnostics:
             "mcse_quantile30": lambda x: mcse(x, method="quantile", prob=0.3),
         }
         results = {}
-        for key, coord_dict, vals in xarray_var_iter(posterior.posterior, combined=True):
+        for key, coord_dict, _, vals in xarray_var_iter(posterior.posterior, combined=True):
             if coord_dict:
                 key = key + ".{}".format(list(coord_dict.values())[0] + 1)
             results[key] = {func_name: func(vals) for func_name, func in funcs.items()}